recommender/requirements.txt
# Đặt numpy và scikit-surprise lên đầu
numpy==1.24.4
scikit-surprise==1.1.3

fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.8.2
SQLAlchemy==1.4.22
psycopg2-binary==2.9.1
confluent-kafka==1.7.0
aiokafka==0.12.0
asyncpg
python-dotenv
pandas 
joblib 
scikit-learn